+++ activerecord/lib/active_record/connection_adapters/connection_specification.rb
@@ -66,6 +66,7 @@ module ActiveRecord
           config = URI.parse url
           adapter = config.scheme
           adapter = "postgresql" if adapter == "postgres"
+          adapter = "mysql2"     if adapter == "mysql"
           spec = { :adapter  => adapter,
                    :username => config.user,
                    :password => config.password,